Add files via upload
While Loops in Java
In programming languages, loops are used to execute a particular statement/set of instructions again and again.
The execution of the loop starts when some conditions become true.
For example, print 1 to 1000, print multiplication table of 7, etc.
Loops make it easy for us to tell the computer that a given set of instructions need to be executed repeatedly.
Types of Loops :
Primarily, there are three types of loops in Java:

While loop
do-while loop
for loop


While loops :
The while loop in Java is used when we need to execute a block of code again and again based on a given boolean condition.
Use a while loop if the exact number of iterations is not known.
If the condition never becomes false, the while loop keeps getting executed. Such a loop is known as an infinite loop.