Add files via upload
For loop:
For loop in java is used to iterate a block of code multiple times.
Use for loop only when the exact number of iterations needed is already known to you. 
Initializer: Initializes the value of a variable. This part is executed only once.
check_bool_expression: The code inside the for loop is executed only when this condition returns true.
update: Updates the value of the initial variable.