Add files via upload
 Methods in Java
Sometimes our program grows in size, and we want to separate the logic of the main method from the other methods.
For instance, if we calculate the average of a number pair 5 times, we can use methods to avoid repeating the logic. [DRY – Don’t Repeat Yourself]
Void return type : 
When we don’t want our method to return anything, we use void as the return type.

Static keyword : 
The static keyword is used to associate a method of a given class with the class rather than the object.
You can call a static method without creating an instance of the class.
In Java, the main() method is static, so that JVM can call the main() method directly without allocating any extra memory for object creation. 
All the objects share the static method in a class.