Add files via upload
Method Overloading in Java
In Java, it is possible for a class to contain two or more methods with the same name but with different parameters. Such methods are called Overloaded methods.
Method overloading is used to increase the readability of the program.
Ways to perform method overloading : 
In Java, method overloading can be performed by two ways listed below :

By changing the return type of the different methods
By changing the number of arguments accepted by the method
Now, let's have an example to understand the above ways of method overloading :

By changing the return type :
In the below example, we've created a class named calculate.
In the calculate class, we've two methods with the same name i.e. multiply
These two methods are overloaded because they have the same name but their return is different.
The return type of 1st method is int while the return type of the other method is double.

By changing the number of arguments passed :
Again, we've created two methods with the same name i.e., multiply
The return type of both the methods is int.
But, the first method 2 arguments and the other method accepts 3 arguments.

Note: Method overloading cannot be performed by changing the return type of methods.