Add files via upload
Dynamic Method Dispatch in Java
Dynamic method dispatch is also known as run time polymorphism.
It is the process through which a call to an overridden method is resolved at runtime.
This technique is used to resolve a call to an overridden method at runtime rather than compile time.
To properly understand Dynamic method dispatch in Java, it is important to understand the concept of upcasting because dynamic method dispatch is based on upcasting.
Upcasting :
It is a technique in which a superclass reference variable refers to the object of the subclass.
Example :
class Animal{}  
class Dog extends Animal{}  
Animal a=new Dog();//upcasting  
In the above example, we've created two classes, named Animal(superclass) & Dog(subclass). While creating the object 'a', we've taken the reference variable of the parent class(Animal), and the object created is of child class(Dog).